+++ pkg/core/interfaces.go
@@ -15,7 +15,6 @@ type SessionResettable interface {
 
 // Borderrer interface for triggering border announcements.
 type Borderrer interface {
-	PlayBorder(ctx context.Context, from, to string, tel *sim.Telemetry) bool
 	Heartbeat(ctx context.Context, tel *sim.Telemetry)
 }